requirements.txt
prometheus-client
RPi.GPIO